refactor: update import paths for product sorting utilities
